requirements.txt
# ===================================
# NetGuardian Pro - Complete Requirements
# ===================================
# Install: pip install -r requirements.txt
# Python: 3.10+

# === Server (FastAPI) ===
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0

# === Database ===
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# === Data Processing ===
pandas>=2.0.0
numpy>=1.24.0

# === Machine Learning ===
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# === Client (Packet Capture) ===
scapy>=2.5.0
grpcio>=1.50.0
grpcio-tools>=1.50.0

# === HTTP Requests ===
requests>=2.31.0

# === Attack Tools (Optional) ===
paramiko>=3.3.0  # SSH brute force simulation

# === Testing & Linting ===
pytest>=7.4.0
flake8>=6.1.0
